Refactor list-files
